+++ packages/kit-bg/src/states/jotai/atomNames.ts
@@ -1,0 +1,3 @@
+export enum EAtomNames {
+  demoPriceAtom = 'demoPriceAtom',
+}
