+++ packages/kit-bg/src/states/jotai/atoms/index.ts
@@ -1,13 +1,1 @@
-import * as demo from './demo';
-
-const allAtoms = {
-  ...demo,
-};
-
-console.log(Object.keys(allAtoms));
-if (process.env.NODE_ENV !== 'production') {
-  // @ts-ignore
-  global.$$allAtoms = allAtoms;
-}
-
-export default allAtoms;
+export * from './demo';
