+++ packages/kit-bg/src/states/jotai/jotaiInit.ts
@@ -1,0 +1,5 @@
+import * as allAtoms from './atoms';
+
+export function jotaiInit() {
+  console.log('allAtoms : ', allAtoms);
+}
