+++ packages/kit-bg/src/states/jotai/jotaiStorage.ts
@@ -1,0 +1,142 @@
+/* eslint-disable camelcase */
+import { atom } from 'jotai';
+
+import appStorage from '@onekeyhq/shared/src/storage/appStorage';
+
+import { RESET } from './types';
+
+import type {
+  AsyncStorage,
+  SetStateActionWithReset,
+  SyncStorage,
+  Unsubscribe,
+  WritableAtom,
+} from './types';
+
+class JotaiStorage implements AsyncStorage<any> {
+  async getItem(key: string, initialValue: any): Promise<any> {
+    const r = await appStorage.getItem(key);
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-return
+    return r ?? initialValue;
+  }
+
+  async setItem(key: string, newValue: any): Promise<void> {
+    await appStorage.setItem(key, newValue);
+  }
+
+  async removeItem(key: string): Promise<void> {
+    await appStorage.removeItem(key);
+  }
+
+  subscribe = undefined;
+}
+
+const onekeyJotaiStorage = new JotaiStorage();
+
+export function atomWithStorage<Value>(
+  storageName: string,
+  initialValue: Value,
+  storage: AsyncStorage<Value>,
+  unstable_options?: { unstable_getOnInit?: boolean },
+): WritableAtom<
+  Value | Promise<Value>,
+  [SetStateActionWithReset<Value | Promise<Value>>],
+  Promise<void>
+>;
+
+export function atomWithStorage<Value>(
+  storageName: string,
+  initialValue: Value,
+  storage?: SyncStorage<Value>,
+  unstable_options?: { unstable_getOnInit?: boolean },
+): WritableAtom<Value, [SetStateActionWithReset<Value>], void>;
+
+// TODO rename to atomPro
+// - support async storage
+// - support storage ready check (apply to raw atom and computed atom)
+// - support Ext ui & bg sync
+export function atomWithStorage<Value>(
+  storageName: string,
+  initialValue: Value,
+  storage: AsyncStorage<Value> | SyncStorage<Value> = onekeyJotaiStorage,
+  unstable_options?: { unstable_getOnInit?: boolean },
+): any {
+  let updateAtomStorageReady: ((v: boolean) => void) | undefined;
+  const key = `global_states:${storageName}`;
+  const getOnInit = unstable_options?.unstable_getOnInit ?? false;
+  const baseAtom = atom(
+    getOnInit
+      ? (storage.getItem(key, initialValue) as Value | Promise<Value>)
+      : initialValue,
+  );
+
+  if (process.env.NODE_ENV !== 'production') {
+    baseAtom.debugPrivate = true;
+  }
+
+  // TODO onMount not trigger when UI not reference this atom
+  // TODO computed atom storage ready not working
+  baseAtom.onMount = (async (setAtom: (value: Value) => void) => {
+    if (!getOnInit) {
+      const val = (await storage.getItem(key, initialValue)) as Value;
+      setAtom(val);
+      if (!updateAtomStorageReady) {
+        throw new Error('updateAtomStorageReady is undefined');
+      }
+      updateAtomStorageReady(true);
+    }
+    let unsub: Unsubscribe | undefined;
+    if (storage.subscribe) {
+      unsub = storage.subscribe(key, setAtom, initialValue);
+    }
+    return unsub;
+  }) as any;
+
+  const anAtom = atom(
+    (get) => get(baseAtom),
+    (get, set, update: SetStateActionWithReset<Value | Promise<Value>>) => {
+      const nextValue =
+        typeof update === 'function'
+          ? (
+              update as (
+                prev: Value | Promise<Value>,
+              ) => Value | Promise<Value> | typeof RESET
+            )(get(baseAtom))
+          : update;
+      if (nextValue === RESET) {
+        set(baseAtom, initialValue);
+        return storage.removeItem(key);
+      }
+      if (nextValue instanceof Promise) {
+        return nextValue.then((resolvedValue) => {
+          set(baseAtom, resolvedValue);
+          return storage.setItem(key, resolvedValue);
+        });
+      }
+      set(baseAtom, nextValue);
+      return storage.setItem(key, nextValue);
+    },
+  );
+
+  const storageReady = new Promise<boolean>((resolve) => {
+    updateAtomStorageReady = resolve;
+    if (process.env.NODE_ENV !== 'production') {
+      // @ts-ignore
+      global[`$$updateAtomStorageReady_${storageName}`] = resolve;
+    }
+  });
+  // @ts-ignore
+  anAtom.storageReady = storageReady;
+
+  return anAtom;
+}
+
+// @ts-ignore
+global.$globalStatsStorageReadyResolve = (v: boolean) => v;
+// @ts-ignore
+global.$globalStatsStorageReady = new Promise<boolean>(
+  // @ts-ignore
+  (resolve) => (global.$globalStatsStorageReadyResolve = resolve),
+);
+
+function jotaiStorageInit() {}
