+++ packages/kit/src/hooks/useAppNavigation.ts
@@ -62,9 +62,9 @@ function useAppNavigation<
   const iosHeaderStyle = Platform.select<StackNavigationOptions>({
     ios: {
       headerStyle: {
-        backgroundColor: bgColor as string,
+        backgroundColor: bgColor,
       },
-      headerTintColor: titleColor as string,
+      headerTintColor: titleColor,
     },
   });
